INIT: split out methods
