+++ src/md_html/exceptions.py
@@ -1,0 +1,15 @@
+"""Custom exceptions for the package."""
+
+
+class HTMLParseError(Exception):
+    """Exception raised during the parsing of markdown text.
+
+    Attributes:
+        message -- explanation of the error
+
+    """
+
+    def __init__(self, message: str) -> None:
+        """Instantiate the HTMLParseError exception."""
+        self.message = message
+        super().__init__(self.message)
