+++ src/md_html/utils.py
@@ -1,0 +1,9 @@
+"""Utility methods the package."""
+
+from enum import StrEnum
+
+
+class Elements(StrEnum):
+    """HTML elements lookup."""
+
+    paragraph = "p"
